backend/requirements.txt
Flask==3.1.3
Flask-CORS==6.0.2
psycopg2-binary==2.9.11
BeautifulSoup4==4.14.3
python-dotenv==1.2.2
Werkzeug==3.1.6
gunicorn==21.2.0